fix(db): referral_id BIGINT + split migration 015 into two transactions
Migration 016: users.referral_id stores referrer_tg_id (BIGINT, like
users.tg_id) but the column was INTEGER — for any referrer with tg_id >
2^31-1 the atomic UPDATE in _attach_referral_if_any raised
`integer out of range` (uncaught → 500 on claim). Cast to BIGINT,
guarded by data_type check; rollback provided but lossy and not
recommended. schema_fixed.sql seeds the column as BIGINT for fresh DBs.

Migration 015: split the single BEGIN/COMMIT into two independent
transactions. Previously dedup, UNIQUE(referred_tg_id) and the
text→DECIMAL cast ran together — if the cast hit a non-numeric
reward_value (junk from old rows) it rolled back the dedup and the
UNIQUE constraint too, so idempotency never landed and every deploy
failed again. tx1: dedup + UNIQUE (dedup must precede UNIQUE);
tx2: clean non-numeric reward_value BEFORE the cast, then ALTER —
a cast failure no longer breaks the UNIQUE constraint.

+++ db/schema_fixed.sql
@@ -77,14 +77,16 @@ CREATE TABLE IF NOT EXISTS users (
     -- См. models/servers/server.py:get_env_server.
 );
 
--- Add missing referral_id (migration 008)
+-- Add missing referral_id (migration 008; BIGINT per migration 016 —
+-- хранит referrer_tg_id, который BIGINT; INTEGER вызывал `integer out of range`
+-- для tg_id > 2^31-1).
 DO $$
 BEGIN
     IF NOT EXISTS (
         SELECT 1 FROM information_schema.columns
         WHERE table_name = 'users' AND column_name = 'referral_id'
     ) THEN
-        ALTER TABLE users ADD COLUMN referral_id INTEGER;
+        ALTER TABLE users ADD COLUMN referral_id BIGINT;
     END IF;
 END $$;
